refactor (useFetch): move url logic to custom hook
